+++ src/hosting/gitea_test.go
@@ -76,7 +76,7 @@ func TestGitea(t *testing.T) {
 			assert.NoError(t, err)
 			assert.True(t, prInfo.CanMergeWithAPI)
 			assert.Equal(t, "my title (#1)", prInfo.DefaultCommitMessage)
-			assert.Equal(t, int64(1), prInfo.PullRequestNumber)
+			assert.Equal(t, 1, prInfo.PullRequestNumber)
 		})
 
 		t.Run("empty Git token", func(t *testing.T) {
